+++ packages/frontend/src/providers/DashboardProvider.tsx
@@ -4,16 +4,15 @@ import {
     compressDashboardFiltersToParam,
     convertDashboardFiltersParamToDashboardFilters,
     Dashboard,
-    DashboardAvailableFilters,
     DashboardFilterRule,
     DashboardFilters,
     fieldId,
     FilterableField,
     isDashboardChartTileType,
+    SavedChartsInfoForDashboardAvailableFilters,
     SortField,
 } from '@lightdash/common';
 import { min } from 'lodash-es';
-import uniqBy from 'lodash-es/uniqBy';
 import React, {
     Dispatch,
     SetStateAction,
@@ -45,7 +44,6 @@ const emptyFilters: DashboardFilters = {
 type DashboardContext = {
     dashboard: Dashboard | undefined;
     dashboardError: ApiError | null;
-    fieldsWithSuggestions: FieldsWithSuggestions;
     dashboardTiles: Dashboard['tiles'] | undefined;
     setDashboardTiles: Dispatch<SetStateAction<Dashboard['tiles'] | undefined>>;
     haveTilesChanged: boolean;
@@ -76,14 +74,13 @@ type DashboardContext = {
     ) => void;
     haveFiltersChanged: boolean;
     setHaveFiltersChanged: Dispatch<SetStateAction<boolean>>;
-    addSuggestions: (newSuggestionsMap: Record<string, string[]>) => void;
     addResultsCacheTime: (cacheMetadata: CacheMetadata) => void;
     oldestCacheTime: Date | undefined;
     invalidateCache: boolean | undefined;
     clearCacheAndFetch: () => void;
+    fieldsWithSuggestions: FieldsWithSuggestions;
     allFilterableFields: FilterableField[] | undefined;
-    filterableFieldsBySavedQueryUuid: DashboardAvailableFilters | undefined;
-    filterableFieldsByTileUuid: DashboardAvailableFilters | undefined;
+    filterableFieldsByTileUuid: Record<string, FilterableField[]> | undefined;
     hasChartTiles: boolean;
     chartSort: Record<string, SortField[]>;
     setChartSort: (sort: Record<string, SortField[]>) => void;
@@ -104,8 +101,6 @@ export const DashboardProvider: React.FC = ({ children }) => {
     const [dashboardTiles, setDashboardTiles] = useState<Dashboard['tiles']>();
 
     const [haveTilesChanged, setHaveTilesChanged] = useState<boolean>(false);
-    const [fieldsWithSuggestions, setFieldsWithSuggestions] =
-        useState<FieldsWithSuggestions>({});
     const [dashboardTemporaryFilters, setDashboardTemporaryFilters] =
         useState<DashboardFilters>(emptyFilters);
     const [dashboardFilters, setDashboardFilters] =
@@ -123,16 +118,25 @@ export const DashboardProvider: React.FC = ({ children }) => {
     const {
         overridesForSavedDashboardFilters,
         addSavedFilterOverride,
-
         removeSavedFilterOverride,
     } = useSavedDashboardFiltersOverrides();
 
-    const tileSavedChartUuids = useMemo(
+    const savedChartUuidsAndTileUuids = useMemo(
         () =>
             dashboardTiles
                 ?.filter(isDashboardChartTileType)
-                .map((tile) => tile.properties.savedChartUuid)
-                .filter((uuid): uuid is string => !!uuid),
+                .reduce<SavedChartsInfoForDashboardAvailableFilters>(
+                    (acc, tile) => {
+                        if (tile.properties.savedChartUuid) {
+                            acc.push({
+                                tileUuid: tile.uuid,
+                                savedChartUuid: tile.properties.savedChartUuid,
+                            });
+                        }
+                        return acc;
+                    },
+                    [],
+                ),
         [dashboardTiles],
     );
 
@@ -251,38 +255,52 @@ export const DashboardProvider: React.FC = ({ children }) => {
     const {
         isLoading: isLoadingDashboardFilters,
         isFetching: isFetchingDashboardFilters,
-        data: filterableFieldsBySavedQueryUuid,
-    } = useDashboardsAvailableFilters(tileSavedChartUuids ?? []);
+        data: dashboardAvailableFiltersData,
+    } = useDashboardsAvailableFilters(savedChartUuidsAndTileUuids ?? []);
 
     const filterableFieldsByTileUuid = useMemo(() => {
-        if (!dashboard || !dashboardTiles || !filterableFieldsBySavedQueryUuid)
+        if (!dashboard || !dashboardTiles || !dashboardAvailableFiltersData)
             return;
 
-        return dashboardTiles
-            .filter(isDashboardChartTileType)
-            .reduce<DashboardAvailableFilters>((acc, tile) => {
-                const savedChartUuid = tile.properties.savedChartUuid;
-                if (!savedChartUuid) return acc;
+        const filterFieldsMapping = savedChartUuidsAndTileUuids?.reduce<
+            Record<string, FilterableField[]>
+        >((acc, { tileUuid }) => {
+            const filterFields =
+                dashboardAvailableFiltersData.savedQueryFilters[tileUuid]?.map(
+                    (index) =>
+                        dashboardAvailableFiltersData.allFilterableFields[
+                            index
+                        ],
+                );
 
-                return {
-                    ...acc,
-                    [tile.uuid]:
-                        filterableFieldsBySavedQueryUuid[savedChartUuid],
-                };
-            }, {});
-    }, [dashboard, dashboardTiles, filterableFieldsBySavedQueryUuid]);
+            if (filterFields) {
+                acc[tileUuid] = filterFields;
+            }
 
-    const allFilterableFields = useMemo(() => {
-        if (isLoadingDashboardFilters || !filterableFieldsBySavedQueryUuid)
-            return;
+            return acc;
+        }, {});
 
-        const allFilters = Object.values(
-            filterableFieldsBySavedQueryUuid,
-        ).flat();
-        if (allFilters.length === 0) return;
+        return filterFieldsMapping;
+    }, [
+        dashboard,
+        dashboardTiles,
+        dashboardAvailableFiltersData,
+        savedChartUuidsAndTileUuids,
+    ]);
 
-        return uniqBy(allFilters, (f) => fieldId(f));
-    }, [isLoadingDashboardFilters, filterableFieldsBySavedQueryUuid]);
+    const fieldsWithSuggestions = useMemo(() => {
+        return dashboardAvailableFiltersData &&
+            dashboardAvailableFiltersData.allFilterableFields &&
+            dashboardAvailableFiltersData.allFilterableFields.length > 0
+            ? dashboardAvailableFiltersData.allFilterableFields.reduce<FieldsWithSuggestions>(
+                  (sum, field) => ({
+                      ...sum,
+                      [fieldId(field)]: field,
+                  }),
+                  {},
+              )
+            : {};
+    }, [dashboardAvailableFiltersData]);
 
     const allFilters = useMemo(() => {
         return {
@@ -305,24 +323,6 @@ export const DashboardProvider: React.FC = ({ children }) => {
             ),
         [dashboardTiles],
     );
-
-    useEffect(() => {
-        if (allFilterableFields && allFilterableFields.length > 0) {
-            setFieldsWithSuggestions((prev) => {
-                return allFilterableFields.reduce<FieldsWithSuggestions>(
-                    (sum, field) => ({
-                        ...sum,
-                        [fieldId(field)]: {
-                            ...field,
-                            suggestions:
-                                prev[fieldId(field)]?.suggestions || [],
-                        },
-                    }),
-                    {},
-                );
-            });
-        }
-    }, [allFilterableFields]);
 
     const addDimensionDashboardFilter = useCallback(
         (filter: DashboardFilterRule, isTemporary: boolean) => {
@@ -422,24 +422,6 @@ export const DashboardProvider: React.FC = ({ children }) => {
         },
         [removeSavedFilterOverride],
     );
-    const addSuggestions = useCallback(
-        (newSuggestionsMap: Record<string, string[]>) => {
-            setFieldsWithSuggestions((prev) => {
-                return Object.entries(prev).reduce<FieldsWithSuggestions>(
-                    (sum, [key, field]) => {
-                        const currentSuggestions = field?.suggestions || [];
-                        const newSuggestions = newSuggestionsMap[key] || [];
-                        const suggestions = Array.from(
-                            new Set([...currentSuggestions, ...newSuggestions]),
-                        ).sort((a, b) => a.localeCompare(b));
-                        return { ...sum, [key]: { ...field, suggestions } };
-                    },
-                    {},
-                );
-            });
-        },
-        [],
-    );
 
     const addResultsCacheTime = useCallback((cacheMetadata: CacheMetadata) => {
         if (cacheMetadata.cacheHit && cacheMetadata.cacheUpdatedTime) {
@@ -466,7 +448,6 @@ export const DashboardProvider: React.FC = ({ children }) => {
     const value = {
         dashboard,
         dashboardError,
-        fieldsWithSuggestions,
         dashboardTiles,
         setDashboardTiles,
         haveTilesChanged,
@@ -481,13 +462,12 @@ export const DashboardProvider: React.FC = ({ children }) => {
         setDashboardFilters,
         haveFiltersChanged,
         setHaveFiltersChanged,
-        addSuggestions,
         addResultsCacheTime,
         oldestCacheTime,
         invalidateCache,
         clearCacheAndFetch,
-        allFilterableFields,
-        filterableFieldsBySavedQueryUuid,
+        fieldsWithSuggestions,
+        allFilterableFields: dashboardAvailableFiltersData?.allFilterableFields,
         isLoadingDashboardFilters,
         isFetchingDashboardFilters,
         filterableFieldsByTileUuid,