+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.863.12](https://github.com/lightdash/lightdash/compare/0.863.11...0.863.12) (2023-11-15)
+
+
+### Bug Fixes
+
+* improve available filters and metric query suggestions ([#7942](https://github.com/lightdash/lightdash/issues/7942)) ([3315c52](https://github.com/lightdash/lightdash/commit/3315c526fd06450a7b632bc8958f835095dcd228))
+
+## [0.863.11](https://github.com/lightdash/lightdash/compare/0.863.10...0.863.11) (2023-11-14)
+
+
+### Bug Fixes
+
+* keep google sync dialog footer in view ([#7956](https://github.com/lightdash/lightdash/issues/7956)) ([a23dc99](https://github.com/lightdash/lightdash/commit/a23dc99bb5738c723b7e4ee4b8d23b50a1a10824))
+
 ## [0.863.10](https://github.com/lightdash/lightdash/compare/0.863.9...0.863.10) (2023-11-14)